+++ db/migrate/20250617081408_add_requirements_linked_to_pfi_to_facilities_management_rm6232_procurements.rb
@@ -1,0 +1,5 @@
+class AddRequirementsLinkedToPfiToFacilitiesManagementRM6232Procurements < ActiveRecord::Migration[8.0]
+  def change
+    add_column :facilities_management_rm6232_procurements, :requirements_linked_to_pfi, :boolean
+  end
+end